Split out classes out of InstanceStackChunkKlass files
