+++ src/hotspot/share/oops/instanceStackChunkKlass.hpp
@@ -29,6 +29,7 @@
 #include "oops/instanceKlass.hpp"
 #include "oops/stackChunkOop.hpp"
 #include "runtime/handles.hpp"
+#include "runtime/stackChunkFrameStream.hpp"
 #include "utilities/macros.hpp"
 
 class ClassFileParser;
@@ -36,8 +37,6 @@ class frame;
 class ImmutableOopMap;
 class VMRegImpl;
 typedef VMRegImpl* VMReg;
-
-template <chunk_frames = chunk_frames::MIXED> class StackChunkFrameStream;
 
 // An InstanceStackChunkKlass is a specialization of the InstanceKlass.
 // It has a header containing metadata, and a blob containing a stack segment
@@ -107,7 +106,6 @@ private:
   friend class InstanceKlass;
   friend class stackChunkOopDesc;
   friend class Continuations;
-  template <chunk_frames frames> friend class StackChunkFrameStream;
   friend class FixChunkIterateStackClosure;
   friend class MarkMethodsStackClosure;
   friend class OopOopIterateStackClosure;
@@ -120,8 +118,6 @@ private:
   static int _offset_of_stack;
 
   InstanceStackChunkKlass(const ClassFileParser& parser);
-  static inline int metadata_words(); // size, in words, of frame metadata (e.g. pc and link)
-  static inline int align_wiggle();   // size, in words, of maximum shift in frame position due to alignment
 
 public:
   InstanceStackChunkKlass() { assert(DumpSharedSpaces || UseSharedSpaces, "only for CDS"); }
@@ -136,6 +132,11 @@ public:
   static inline size_t bitmap_size(size_t stack_size_in_words); // in words
   // the *last* bit in the bitmap corresponds to the last word in the stack; this returns the bit index corresponding to the first word
   static inline BitMap::idx_t bit_offset(size_t stack_size_in_words);
+
+  // size of frame metadata (e.g. pc and link) - in words
+  static inline int metadata_words();
+  // size of maximum shift in frame position due to alignment - in words
+  static inline int align_wiggle();
 
   // Returns the size of the instance including the stack data.
   virtual size_t oop_size(oop obj) const override;
@@ -238,100 +239,6 @@ private:
 
   template <copy_alignment alignment> inline static void copy_from_stack_to_chunk(void* from, void* to, size_t size);
   template <copy_alignment alignment> inline static void copy_from_chunk_to_stack(void* from, void* to, size_t size);
-};
-
-template <chunk_frames frame_kind>
-class StackChunkFrameStream : public StackObj {
-private:
-  intptr_t* _end;
-  intptr_t* _sp;
-  intptr_t* _unextended_sp; // used only when mixed
-  CodeBlob* _cb;
-  mutable const ImmutableOopMap* _oopmap;
-
-#ifndef PRODUCT
-  stackChunkOop _chunk;
-  int _index;
-#endif
-
-#ifdef ASSERT
-  int _has_stub;
-#endif
-
-public:
-  StackChunkFrameStream() { NOT_PRODUCT(_chunk = nullptr; _index = -1;) DEBUG_ONLY(_has_stub = false;) }
-  inline StackChunkFrameStream(stackChunkOop chunk, bool gc = false);
-  inline StackChunkFrameStream(stackChunkOop chunk, const frame& f);
-
-  bool is_done() const { return _sp >= _end; }
-  bool is_last() const { return next_sp() >= _end; }
-
-  intptr_t* end() { return _end; }
-  void set_end(intptr_t* end) { _end = end; }
-
-  // Query
-  intptr_t* end() const { return _end; }
-
-  intptr_t*        sp() const  { return _sp; }
-  inline address   pc() const  { return get_pc(); }
-  inline intptr_t* fp() const;
-  inline intptr_t* unextended_sp() const { return frame_kind == chunk_frames::MIXED ? _unextended_sp : _sp; }
-  NOT_PRODUCT(int index() { return _index; })
-  inline address orig_pc() const;
-
-  inline bool is_interpreted() const;
-  inline bool is_stub() const;
-  inline bool is_compiled() const;
-  CodeBlob* cb() const { return _cb; }
-  const ImmutableOopMap* oopmap() const { if (_oopmap == NULL) get_oopmap(); return _oopmap; }
-  inline int frame_size() const;
-  inline int stack_argsize() const;
-  inline int num_oops() const;
-
-  inline void initialize_register_map(RegisterMap* map);
-  template <typename RegisterMapT> inline void next(RegisterMapT* map);
-
-  template <typename RegisterMapT> inline void update_reg_map(RegisterMapT* map);
-
-  void handle_deopted() const;
-
-  inline int to_offset(stackChunkOop chunk) const { assert(!is_done(), ""); return _sp - chunk->start_address(); }
-
-  inline frame to_frame() const;
-
-#ifdef ASSERT
-  bool is_in_frame(void* p) const;
-  bool is_deoptimized() const;
-  template <typename RegisterMapT> bool is_in_oops(void* p, const RegisterMapT* map) const;
-#endif
-
-  void print_on(outputStream* st) const PRODUCT_RETURN;
-
- private:
-  inline address get_pc() const;
-  inline void get_cb();
-
-  inline intptr_t* next_sp() const;
-  inline int interpreter_frame_size() const;
-  inline int interpreter_frame_num_oops() const;
-  inline int interpreter_frame_stack_argsize() const;
-  inline void next_for_interpreter_frame();
-  inline intptr_t* next_sp_for_interpreter_frame() const;
-  inline intptr_t* unextended_sp_for_interpreter_frame() const;
-  inline intptr_t* derelativize(int offset) const;
-  inline void get_oopmap() const;
-  inline void get_oopmap(address pc, int oopmap_slot) const;
-
-  template <typename RegisterMapT> inline void update_reg_map_pd(RegisterMapT* map);
-
-  template <typename RegisterMapT>
-  inline void* reg_to_loc(VMReg reg, const RegisterMapT* map) const;
-
-public:
-  template <class OopClosureType, class RegisterMapT>
-  inline void iterate_oops(OopClosureType* closure, const RegisterMapT* map) const;
-  template <class DerivedOopClosureType, class RegisterMapT>
-  inline void iterate_derived_pointers(DerivedOopClosureType* closure, const RegisterMapT* map) const;
 };
 
 #endif // SHARE_OOPS_INSTANCESTACKCHUNKKLASS_HPP
